requirements.txt
# Telegram Bot Framework
python-telegram-bot>=20.7

# Google Gemini AI (new package)
google-genai>=0.1.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Utilities
python-dotenv>=1.0.0
pillow>=10.1.0

# Async support
aiohttp>=3.9.1
Pillow>=10.0.0
